// src/app/models/PhieuKiemNghiemChatLuongHang.ts
export class PhieuKiemNghiemChatLuongHang {
    ddiemBquan: string;
    hthucBquan: string;
    id: number;
    maHhoa: string;
    maKho: string;
    maNgan: string;
    ngayKnghiem: string;
    ngayLayMau: string;
    ngayNhapDay: string;
    sluongBquan: number;
    soBbanKthucNhap: string;
    soPhieu: string;
    tenHhoa: string;
    tenKho: string;
    tenNgan: string;
    trangThai: string;
    tenDonVi: string;
    maDonVi: string;
    constructor() {
    }
}